Initial Commit
Very rough draft with html and css to start. Added some images for the background in a folder couldn't decide which to use. We can use another.

To preview the site, drag and drop the "base.html" to your browser. Anytime you make a change, make sure to save the file then refresh the browser to see your changes.